Create DAY 14-Sorting An Array
Arrays.sort() - method consists of two variations one in which we do not pass any arguments where it sort down the complete array be it integer array or character array but if we are supposed to sort a specific part using this method of Arrays class then we overload it and pass the starting and last index to the array.

Parameters: It takes three parameters as can be perceived from the syntax which are as follows:

The array to be sorted
The index of the first element, inclusive, to be sorted (Referred to as from_index)
The index of the last element, exclusive, to be sorted (Referred to as last_index)

Using the reverse order method (as in the program): This method will sort the array in descending. Java Collections class also provides the reverseOrder() method to sort the array in reverse-lexicographic order.  It does not parse any parameter because static method, so we can invoke it directly by using the class name. it will sort arrays in ascending order by the sort() method after that the reverse order() method will give us the natural ordering and we will get the sorted array in descending order.