Create DAY 22- Ways of Printing 2D array
Method 1 (Simple Traversal) 
We have got the input of the number of rows and columns from the user.
Traversed it till the length of array with the size of (i rows * j columns) printing each element.

Method 2 (Using for-each loop) 
 Instead of simple for loops, we used for each loop here. 

Method 3 ( Using Arrays.toString()) 
Arrays.toString(row) converts the complete row is converted as a string and then each row is printed in a separate line. 

Method 4 ( Using Arrays.deepToString())
Arrays.deepToString(int[][]) converts  2D array to string in a single step.