Create DAY 23- Sorting a user's choice column in 2D Array
We are taking an input of a 2D array of order N X M and a column number K ( 1<=K<=m). Our task is to sort the 2D array according to values in Column K.
In the below example, a 2D array of 3x3 order and a column number 3(1<=3<=3) is taken. Our task is to sort the below 2D array according to values in column 3 

Example:
Input : If our 2D array is given as (Order 3X3) 
        39 27 11 
        10 93 91  
        54 78 56  
        24 64 20 
        Sorting it by values in column 3 
Output :
         39 27 11  
         24 64 20 
         54 78 56 
         10 93 91 
* A comparator interface has been used in the program. 
A comparison function, which imposes a total ordering on some collection of objects.
 Comparators can be passed to a sort method (such as Collections.sort or Arrays.sort) to allow precise control over the sort order.
 Comparators can also be used to control the order of certain data structures (such as sorted sets or sorted maps), or to provide an ordering for collections of objects that don't have a natural ordering.