Create DAY 31- Replace A character From String
Input-
Enter a string: 
Java is Object Oriented Language
enter the index at which you want the character from: 
11
Enter the character you want to replace from that index: 
J

Output-
Original String = Java is Object Oriented Language
Modified String Using String Class= Java is ObjJct Oriented Language
Modified String Using String Builder= Java is ObjJct Oriented Language
Modified String Using String Buffer= Java is ObjJct Oriented Language

Method 1: Using String Class
There is no predefined method in String Class to replace a specific character in a String, as of now. However, this can be achieved indirectly by constructing a new String with 2 different substrings, one from the beginning till the specific index – 1, the new character at the specific index, and the other from the index + 1 till the end.

Method 2: Using StringBuilder
Unlike String Class, the StringBuilder class is used to represent a mutable string of characters and has a predefined method for change a character at a specific index – setCharAt(). Replace the character at the specific index by calling this method and passing the character and the index as the parameter.

Method 3: Using StringBuffer
Like StringBuilder, the StringBuffer class has a predefined method for this purpose – setCharAt(). Replace the character at the specific index by calling this method and passing the character and the index as the parameter. StringBuffer is thread-safe and can be used in a multi-threaded environment. StringBuilder is faster when compared to StringBuffer, but is not thread-safe.

Implementations of all three above methods have been given in the above program.