Create DAY 37-  Determine the Unicode Code Point at Given Index in String
OUTPUT
Original String : peers
unicode point at 0 = 112
unicode point at 1 = 101
unicode point at 2 = 101
unicode point at 3 = 114
unicode point at 4 = 115

ABOUT-
ASCII is a code that converts the English alphabets to numerics as numeric can convert to the assembly language which our computer understands. For that, we have assigned a number against each character ranging from 0 to 127. Alphabets are case-sensitive lowercase and uppercase are treated differently. Complete ASCII values table can be interpreted but better it is sticking to the table in the illustration below where no need arises to lean complete ASCII values in a table. Extremities are assigned as follows so as order to guess the Unicode value correctly with memorizing the complete table. With this table, one can extract all alphabets Unicode be it uppercase or lowercase.



ILLUSTRATION:

S.No	 Cases	                                Unicode Value
1	      Uppercase starting of alphabets	65
2	      Uppercase ending of alphabets	90
3	      Lowercase starting of alphabets	97
4	Lowercase ending of alphabets	122
From this, one can get Unicode of other values such as B is 66 because lowercase starting is a for which is illustrated from the above table, its 65. For ‘b’ it’s 98. Similarly, for Unicode of ‘G’ is 71, ‘E’ is 69, ‘K’ is 75, and so on.

ABOUT FUNCTION USED-
codePointAt() inbuilt method is used where the user wants to return the character at the specific index. The index refers to character values (Unicode units)  and ranges from 0 to length()-1

Definition: This is an inbuilt function that Returns the character(Unicode point) at the specific index. The index refers to character values (Unicode units) and ranges from 0 to length()-1.

Syntax:
java.lang.String.codePointAt();

Parameter: The index to the character values.

Return Type: This method returns the Unicode value at the specified index. The index refers to char values (Unicode code units) and ranges from 0 to [length()-1]. Simply in layman's language, the code point value of the character at the index.

Implementation: This is an example where there is no involvement of exception with this function.