Create DAY 51-Get first and last elements from ArrayList
OUTPUT-
ArrayList contains: [5, 4, 3, 2, 1]
Minimum Value: 1
Maximum Value: 5

ABOUT-
The last element is at index, size – 1 and the first element is stored at index 0. If we know how to get the size of ArrayList then we can get those two values easily. But remember, that you need to use size() method for ArrayList, not length, length is used to get the length of an array.

Application : The first element is your lowest and the last element is your highest in case of ascending order and opposite then the first element would be the maximum and last element would be the minimum if List is in descending order.