Create DAY 57- Display Dates of a Calendar Year in Different Format
OUTPUT-
2022-12-17

CONCEPT:
 Whenever it comes down to the date and time the primary goal is to Date Class. It is necessary to be imported into our program.

Date Class was introduced in java in JDK version 1.0 but was later improved in JDK 1.1 by introducing Calendar Class. It deprecated most Calendar Class as 4 out of 6 constructors are marked as deprecated and most of the methods inside it. These two JDK are outdated now and hence known as legacy API. Finally, in 2014 in JDK8 new date/Time API was introduced. it includes java.time package which is now the main API for a date, time, instants, and duration.

1. java.time.LocalDate– represents a date (year, month, day (yyyy-MM-dd)).
2. java.time.LocalDateTime– represents both a date and a time (yyyy-MM-dd-HH-mm-ss-ns).
3. java.time.format.DateTimeFormatter– formatter for displaying and parsing date-time objects.

When there is a need to display date and time in a different format, the ofPattern() method is used to accept all sorts of values. There are numerous formats that can be invoked later on can be mixed and match letters to achieve the required pattern.

Here the dealing can be with 3 types of formats:
1. yyyy-MM-dd
2. DD/MM/YY
3. dd MMM yyyy