Create DAY 62- Runtime Error
OUTPUT-
Exception in thread "main" java.lang.ArithmeticException: / by zero
	at Exceptions.RunTimeException.main(RunTimeException.java:14)

-->Run Time Error: 
Run Time errors occur or we can say, are detected during the execution of the program. Sometimes these are discovered when the user enters an invalid data or data which is not relevant. Runtime errors occur when a program does not contain any syntax errors but asks the computer to do something that the computer is unable to reliably do. During compilation, the compiler has no technique to detect these kinds of errors. It is the JVM (Java Virtual Machine) that detects it while the program is running. To handle the error during the run time we can put our error code inside the try block and catch the error inside the catch block. 

For example: if the user inputs a data of string format when the computer is expecting an integer, there will be a runtime error. The program shows Runtime Error caused by dividing by zero