Create DAY 63- Compile Time Error
OUTPUT-
Exception in thread "main" java.lang.Error: Unresolved compilation problem: 
	sum cannot be resolved to a variable

	at Exceptions.CompileTimeError.main(CompileTimeError.java:13)

Compile Time Error: 
Compile Time Errors are those errors which prevent the code from running because of an incorrect syntax such as a missing semicolon at the end of a statement or a missing bracket, class not found, etc. These errors are detected by the java compiler and an error message is displayed on the screen while compiling. Compile Time Errors are sometimes also referred to as Syntax errors. These kind of errors are easy to spot and rectify because the java compiler finds them for you. The compiler will tell you which piece of code in the program got in trouble and its best guess as to what you did wrong. Usually, the compiler indicates the exact line where the error is, or sometimes the line just before it, however, if the problem is with incorrectly nested braces, the actual error may be at the beginning of the block. In effect, syntax errors represent grammatical errors in the use of the programming language. 

The program shows misspelled variable name or method names