Create DAY 66- Unchecked Exception
OUTPUT-
Out of index  please check your code

Unchecked Exceptions:
These types of exceptions occur during the runtime of the program.  These are the exceptions that are not checked at a compiled time by the compiler. In Java exceptions under Error and RuntimeException classes are unchecked exceptions, This Exception occurs due to bad programming. Runtime Exceptions like IndexoutOfBoundException, Nullpointer Exception, etc can be handled with the help of Try-Catch Block

(Array)IndexoutOfBoundException: 
This Exception occurs due to accessing the index greater than and equal to the size of the array length. The program will automatically be terminated after this exception.