Create DAY 69- Array Out Of Bound Exception
OUTPUT-
The catch block with Null Pointer Exception is executed
End of program

ABOUT-
Java has a robust Error Handling Mechanism that lets us handle multiple Exceptions in one try block using different catch blocks.  Catch blocks in java are like if-else statements that will become active when an exception occurs. When an exception occurs, the program compares the exception object generated to the exception specified in the catch blocks. The program checks the first catch block then moves on to other and so-on until the generated exception is matched. If no catch block is matched, the program halts, and an exception is thrown at the console.

After an Exception is generated in the try block, the control immediately shifts to the catch block, and try block will no longer execute.