Create DAY 72-Collection Type(Linked List)
OUTPUT-
[1, 2, 3, 4, 5]
[1, 2, 3, 5]
1 2 3 5 

LinkedList: LinkedList class is an implementation of the LinkedList data structure which is a linear data structure where the elements are not stored in contiguous locations and every element is a separate object with a data part and address part. The elements are linked using pointers and addresses. Each element is known as a node.