Create DAY 73-Collection Type(Vector)
OUTPUT-
[1, 2, 3, 4, 5]
[1, 2, 3, 5]
1 2 3 5 

Vector: A vector provides us with dynamic arrays in Java. Though, it may be slower than standard arrays but can be helpful in programs where lots of manipulation in the array is needed. This is identical to ArrayList in terms of implementation. However, the primary difference between a vector and an ArrayList is that a Vector is synchronized and an ArrayList is non-synchronized.