Create DAY 74-Collections Type(Stack)
OUTPUT-
Java Collections Type Stack 
Java Collections Type 

Stack: Stack class models and implements the Stack data structure. The class is based on the basic principle of last-in-first-out. In addition to the basic push and pop operations, the class provides three more functions of empty, search and peek. The class can also be referred to as the subclass of Vector.