Create DAY 80-Collection Type(HashMap)
OUTPUT-
Value for 1 is Java
1 Java
2 Collections
3 HashMap

HashMap: HashMap provides the basic implementation of the Map interface of Java. It stores the data in (Key, Value) pairs. To access a value in a HashMap, we must know its key. HashMap uses a technique called Hashing. Hashing is a technique of converting a large String to a small String that represents the same String so that the indexing and search operations are faster. HashSet also uses HashMap internally.