Create DAY 81-Check Thread Status
OUTPUT-
true
true
run method 
run method 
sleep
sleep

Thread Class
Thread class in java provides numerous methods that are very essential in order to understand the working of threads as the thread stages are triggered by them. Java multi-threading provides two ways to find with the help of isAlive() and join() method.
One thread gets to know when another thread has ended. 

 isAlive() method
- Basically, this method works internally very closely in parallel to the lifecycle stages of a thread. It tests if this thread is alive. A thread is alive if it has been started and has not yet died. There is a transitional period from when a thread is running to when a thread is not running. 
After the run() method returns, there is a short period of time before the thread stops. If we want to know if the start method of the thread class has been called or if the thread has been terminated, we must use the isAlive() method. This method is used to determine if a thread has been started and has not been terminated.