Create DAY 84-Daemon Thread
OUTPUT-
t1 is Daemon thread
t3 is Daemon thread
t2 is User thread

Daemon Thread-
Daemon thread in Java is a low-priority thread that runs in the background to perform tasks such as garbage collection. Daemon thread in Java is also a service provider thread that provides services to the user thread. Its life depends on the mercy of user threads i.e. when all the user threads die, JVM terminates this thread automatically.

In simple words, we can say that it provides services to user threads for background supporting tasks. It has no role in life other than to serve user threads.

Example of Daemon Thread in Java: Garbage collection in Java (gc), finalizer, etc.

Properties of Java Daemon Thread
1. They can not prevent the JVM from exiting when all the user threads finish their execution.
2. JVM terminates itself when all user threads finish their execution.
3. If JVM finds a running daemon thread, it terminates the thread and, after that, shutdown it. JVM does not care whether the Daemon thread is running or not.
4. It is an utmost low priority thread.

Default Nature of Daemon Thread
By default, the main thread is always non-daemon but for all the remaining threads, daemon nature will be inherited from parent to child. That is, if the parent is Daemon, the child is also a Daemon and if the parent is a non-daemon, then the child is also a non-daemon.

Note: Whenever the last non-daemon thread terminates, all the daemon threads will be terminated automatically.

Methods of Daemon Thread
1. void setDaemon(boolean status): 
This method marks the current thread as a daemon thread or user thread. For example, if I have a user thread tU then tU.setDaemon(true) would make it a Daemon thread. On the other hand, if I have a Daemon thread tD then calling tD.setDaemon(false) would make it a user thread. 

Syntax: 
public final void setDaemon(boolean on)

Parameters:
on: If true, marks this thread as a daemon thread.

Exceptions:
 IllegalThreadStateException: if only this thread is active.
SecurityException: if the current thread cannot modify this thread.

2. boolean isDaemon(): 
This method is used to check that the current thread is a daemon. It returns true if the thread is Daemon. Else, it returns false. 

Syntax: 
public final boolean isDaemon()

Returns: 
This method returns true if this thread is a daemon thread; false otherwise