Create DAY 90-hasNext method
OUTPUT-
true
true
false

hasNext() method
hasNext() method is used to check whether there is any element remaining in the List. This method is a boolean type method that returns only true and false  as it is just used for checking purposes. The hasNext() methods of the Iterator and List Iterator returns true if the collection object over which is used to check during traversal whether the pointing element has the next element. If not it simply returns false. So,

Return Value:
        True  - if iteration has more elements 
        False - if iteration has no more elements

Return type: boolean