Create DAY 91-next method
OUTPUT-
Element1
Element2
Element3

next() method-
If there is an element after where hasNext() has returned false on which some execution is to be performed then this method is used to display that element on which execution is supposed to be carried on with help of this method. The next() methods of the Iterator and List Iterator return the next element of the collection. And if there is a need to remove this element remove() method is used.

Return type: Same as collection such as ArrayList, Linked List, etc.

Return value: The next element in the iteration.

Exception: Throws NoSuchElementException if the iteration has no more elements.