Create DAY 96-Delete a node from the beginning
OUTPUT-
Original List: 
1 2 3 4 
Updated List: 
2 3 4 
Updated List: 
3 4 
Updated List: 
4 
Updated List: 
List is empty

ABOUT-
In this program, we created a singly linked list and delete a node from the beginning of the list. To accomplish this task, we need to make the head pointer point to the immediate next of the initial node which will now become the new head node of the list.

Java program to delete a node from the beginning of the singly linked list
Consider the above example; Node was the head of the list. Make head to point to next node in the list. Now, node 1 will become the new head of the list. Thus, deleting the Node.

Algorithm
1.Create a class Node which has two attributes: data and next. Next is a pointer to the next node in the list.
2.Create another class DeleteStart which has two attributes: head and tail.
addNode() will add a new node to the list:
3.Create a new node.
4.It first checks, whether the head is equal to null which means the list is empty.
5.If the list is empty, both head and tail will point to a newly added node.
6.If the list is not empty, the new node will be added to end of the list such that tail's next will point to a newly added node. This new node will become the new tail of the list.

a. deleteFromStart() will delete a node from the beginning of the list:
1.It first checks whether the head is null (empty list) then, display the message "List is empty" and return.
2.If the list is not empty, it will check whether the list has only one node.
3.If the list has only one node, it will set both head and tail to null.
4.If the list has more than one node then, the head will point to the next node in the list and delete the old head node.

a. display() will display the nodes present in the list:
1.Define a node current which will initially point to the head of the list.
2.Traverse through the list till current points to null.
3.Display each node by making current to point to node next to it in each iteration.